+++ Sources/ClientRuntime/Networking/Streaming/BufferedStream.swift
@@ -7,11 +7,18 @@
 
 import Foundation
 
+/// A `Stream` implementation that buffers data in memory.
+/// The buffer size depends on the amount of data written and read.
+/// Note: This class is thread-safe.
 public class BufferedStream: Stream {
+
+    /// The length of the stream, if known.
     public var length: Int?
 
+    /// The current position in the stream.
     public var position: Data.Index
 
+    /// Whether the stream is empty.
     public var isEmpty: Bool {
         return buffer.isEmpty
     }
@@ -21,6 +28,10 @@ public class BufferedStream: Stream {
     var buffer = Data()
     let lock = NSLock()
 
+    /// Initializes a new `BufferedStream` instance.
+    /// - Parameters:
+    ///   - data: The initial data to buffer.
+    ///   - isClosed: Whether the stream is closed.
     public init(data: Data = .init(), isClosed: Bool = false) {
         self.buffer = data
         self.position = data.startIndex
@@ -28,6 +39,9 @@ public class BufferedStream: Stream {
         self.isClosed = isClosed
     }
 
+    /// Reads up to `count` bytes from the stream.
+    /// - Parameter count: The maximum number of bytes to read.
+    /// - Returns: Data read from the stream, or nil if the stream is closed and no data is available.
     public func read(upToCount count: Int) throws -> Data? {
         lock.withLockingClosure {
             let toRead = min(count, buffer.count)
@@ -50,9 +64,12 @@ public class BufferedStream: Stream {
         }
     }
 
+    /// Reads all data from the stream.
+    /// It uses `read(upToCount:)` to read data in chunks.
+    /// - Returns: Data read from the stream, or nil if the stream is closed and no data is available.
     public func readToEnd() throws -> Data? {
         var data = Data()
-        
+
         while let next = try read(upToCount: Int.max) {
             data.append(next)
         }
@@ -68,12 +85,23 @@ public class BufferedStream: Stream {
         }
     }
 
+    /// Seeks to the specified offset in the stream.
+    /// - Parameter offset: The offset to seek to.
     public func seek(toOffset offset: Int) throws {
-        lock.withLockingClosure {
-            self.position = buffer.startIndex.advanced(by: offset)
+        try lock.withLockingThrowingClosure {
+            let newPosition = buffer.startIndex.advanced(by: offset)
+
+            // make sure the new position is within the bounds of the buffer
+            guard newPosition >= buffer.startIndex && newPosition <= buffer.endIndex else {
+                throw StreamError.invalidOffset("Invalid offset: \(offset)")
+            }
+
+            position = newPosition
         }
     }
 
+    /// Writes the specified data to the stream.
+    /// - Parameter data: The data to write.
     public func write(contentsOf data: Data) throws {
         lock.withLockingClosure {
             // append the data to the buffer
@@ -83,33 +111,10 @@ public class BufferedStream: Stream {
         }
     }
 
+    /// Closes the stream.
     public func close() throws {
         lock.withLockingClosure {
             isClosed = true
         }
     }
 }
-
-extension NSLock {
-     /// Execute a closure while holding the lock
-     /// - Parameter closure: A closure to execute while holding the lock
-     /// - Returns: The return value of the closure
-     func withLockingClosure<T>(closure: () -> T) -> T {
-         lock()
-         defer {
-             unlock()
-         }
-         return closure()
-     }
-
-     /// Execute a throwing closure while holding the lock
-     /// - Parameter closure: A throwing closure to execute while holding the lock
-     /// - Returns: The return value of the closure
-     func withLockingThrowingClosure<T>(closure: () throws -> T) throws -> T {
-         lock()
-         defer {
-             unlock()
-         }
-         return try closure()
-     }
- }
